# Conflicts:
#	health-mall-product-model/pom.xml
#	health-mall-product-remote-api/pom.xml

+++ health-mall-product-model/src/main/java/com/drstrong/health/product/model/entity/chinese/ChineseSkuInfoEntity.java
@@ -8,7 +8,6 @@ import lombok.*;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * <p>
@@ -58,9 +57,9 @@ public class ChineseSkuInfoEntity extends BaseStandardEntity implements Serializ
     private Long storeId;
 
     /**
-     * 价格/克
+     * 价格/克,单位：分
      */
-    private BigDecimal price;
+    private Long price;
 
     /**
      * sku上下架状态；0-未上架，1-已上架
